split nextup into files
