+++ srl/commands/nextup/list_.py
@@ -1,0 +1,26 @@
+from rich.console import Console
+from rich.panel import Panel
+from srl.commands.list_ import format_problem
+from .util import get_next_up_problems
+
+
+def handle_list(args, console: Console):
+    next_up = get_next_up_problems()
+
+    if not next_up:
+        console.print("[yellow]Next Up queue is empty.[/yellow]")
+        return
+
+    lines = []
+    for i, (name, url) in enumerate(next_up, start=1):
+        display = format_problem(name, url)
+        lines.append(f"{i}. {display}")
+
+    console.print(
+        Panel.fit(
+            "\n".join(lines),
+            title=f"[bold cyan]Next Up Problems ({len(next_up)})[/bold cyan]",
+            border_style="cyan",
+            title_align="left",
+        )
+    )
